+++ packages/clerk-js/CHANGELOG.md
@@ -1,5 +1,42 @@
 # Change Log
 
+## 5.65.0
+
+### Minor Changes
+
+- Add handling of new Handshake nonce flow when authenticating requests ([#5865](https://github.com/clerk/javascript/pull/5865)) by [@jacekradko](https://github.com/jacekradko)
+
+- Fix issue where we were not correctly passing the checkoutProps through within the PricingTable component. Removes internal checkoutProps prefix from PricingTableBaseProps. ([#5888](https://github.com/clerk/javascript/pull/5888)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+### Patch Changes
+
+- Added a notice in tooltip when member no has permissions to manage billing for all manager related buttons ([#5852](https://github.com/clerk/javascript/pull/5852)) by [@octoper](https://github.com/octoper)
+
+- Fix and issue after paying with a new card where it was navigationg to existing payment sources and then complete checkout screen ([#5882](https://github.com/clerk/javascript/pull/5882)) by [@octoper](https://github.com/octoper)
+
+- Update profile components plans page heading from `Switch plans` to `Plans` ([#5889](https://github.com/clerk/javascript/pull/5889)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+- Sort payment sources so always default one is on top ([#5875](https://github.com/clerk/javascript/pull/5875)) by [@octoper](https://github.com/octoper)
+
+- Add element descriptor to the text element rendered within UserPreview component. ([#5878](https://github.com/clerk/javascript/pull/5878)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+- Remove no manage billing permissions alert from the of start billing page ([#5873](https://github.com/clerk/javascript/pull/5873)) by [@octoper](https://github.com/octoper)
+
+- Fix mobile `<Drawer />` sizing. ([#5880](https://github.com/clerk/javascript/pull/5880)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+- Add descriptor ids to `UserMembershipList` and `OrganizationSwitcherTrigger` elements to improve styling experience. ([#5881](https://github.com/clerk/javascript/pull/5881)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+- Removes white space from PlanCard when the ctaPosition is top and there are no features ([#5885](https://github.com/clerk/javascript/pull/5885)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+- Fix the "Plan starts at" date when a user downgrades from a paid plan to the free plan ([#5893](https://github.com/clerk/javascript/pull/5893)) by [@nikosdouvlis](https://github.com/nikosdouvlis)
+
+- Add loading state to `<PaymentSources />` component. ([#5879](https://github.com/clerk/javascript/pull/5879)) by [@alexcarpenter](https://github.com/alexcarpenter)
+
+- Updated dependencies [[`d81c1c1`](https://github.com/clerk/javascript/commit/d81c1c168e5fa97793f79917d4981af685321cc4), [`0f5145e`](https://github.com/clerk/javascript/commit/0f5145e164f3d3d5faf57e58162b05e7110d2403), [`afdfd18`](https://github.com/clerk/javascript/commit/afdfd18d645608dec37e52a291a91ba5f42dcbe7), [`b7c51ba`](https://github.com/clerk/javascript/commit/b7c51baac6df1129b468274c9a7f63ca303f16ce), [`437b53b`](https://github.com/clerk/javascript/commit/437b53b67e281d076b5b3f927e11c1d64666d154), [`5217155`](https://github.com/clerk/javascript/commit/52171554250c5c58f4f497b6d3c7416e79ac77da)]:
+  - @clerk/localizations@3.15.2
+  - @clerk/types@4.58.0
+  - @clerk/shared@3.8.2
+
 ## 5.64.2
 
 ### Patch Changes